feat: add model entrypoint
Move models to models.py
Add Washburn entrypoint
Add Washburn-Rideal entrypoint
Make unidirect_analyzer access entrypoint
