+++ doc/source/example.yml
@@ -1,5 +1,6 @@
 data1:
   type: Unidirectional
+  model: Washburn
   path: $WETTINGFRONT_SAMPLES/example.mp4
   output-vid: output/example.mp4
   output-data: output/example.csv
