+++ doc/source/intro.rst
@@ -32,6 +32,7 @@ either as YAML or JSON.
 
    data1:
       type: Unidirectional
+      model: Washburn
       path: your-video.mp4
       output-vid: result.mp4
       output-data: result.csv
