+++ src/wettingfront/__init__.py
@@ -11,11 +11,9 @@ import json
 import logging
 import os
 import sys
-from typing import Callable, Tuple
 
 import numpy as np
 import yaml
-from scipy.optimize import curve_fit, root  # type: ignore[import-untyped]
 
 if sys.version_info < (3, 10):
     from importlib_metadata import entry_points
@@ -27,8 +25,6 @@ else:
 __all__ = [
     "get_sample_path",
     "analyze_files",
-    "fit_washburn",
-    "fit_washburn_rideal",
 ]
 
 
@@ -84,7 +80,7 @@ def analyze_files(*paths: str) -> bool:
     # load analyzers
     ANALYZERS = {}
     for ep in entry_points(group="wettingfront.analyzers"):
-        ANALYZERS[ep.name] = ep.load()
+        ANALYZERS[ep.name] = ep
 
     ok = True
     for path in paths:
@@ -110,7 +106,7 @@ def analyze_files(*paths: str) -> bool:
                 typename = v["type"]
                 analyzer = ANALYZERS.get(typename, None)
                 if analyzer is not None:
-                    analyzer(k, v)
+                    analyzer.load()(k, v)
                 else:
                     logging.error(
                         f"Skipping entry: '{path}::{k}' (unknown type: '{typename}')"
@@ -120,90 +116,6 @@ def analyze_files(*paths: str) -> bool:
                 ok = False
                 continue
     return ok
-
-
-def fit_washburn(t, L) -> Tuple[Callable, Tuple[np.float64, np.float64, np.float64]]:
-    r"""Fit data to Washburn's equation [#f1]_.
-
-    The data are fitted to:
-
-    .. math::
-
-        L = k \sqrt{t}
-
-    where :math:`k` is penetrativity of the liquid.
-
-    Arguments:
-        t (array_like, shape (M,)): Time.
-        L (array_like, shape (M,)): Penetration length.
-
-    Returns:
-        func
-            Washburn equation function f(t, k).
-        (k,)
-            Argument for *func*.
-
-    .. [#f1] Washburn, E. W. (1921). The dynamics of capillary flow.
-             Physical review, 17(3), 273.
-    """
-
-    def func(t, k):
-        return k * np.sqrt(t)
-
-    ret, _ = curve_fit(func, t, L)
-    return func, ret
-
-
-def fit_washburn_rideal(
-    t, z
-) -> Tuple[Callable, Tuple[np.float64, np.float64, np.float64]]:
-    r"""Fit data to Washburn-Rideal equation [#f2]_.
-
-    The data are fitted to:
-
-    .. math::
-
-        t = \frac{\alpha}{2\beta}z^2 - \frac{1}{\alpha}\ln{\frac{\alpha}{\sqrt{\beta}}z}
-
-    where :math:`\alpha` and :math:`\beta` denotes the ratios of viscous drag,
-    surface tension and inertial force [#f3]_.
-
-    Arguments:
-        t (array_like, shape (M,)): Time.
-        z (array_like, shape (M,)): Penetration length.
-
-    Returns:
-        func
-            Washburn-Rideal equation function f(t, alpha, beta).
-        (alpha, beta)
-            Arguments for *func*.
-
-    .. [#f2] Rideal, E. K. (1922). CVIII. On the flow of liquids under capillary
-             pressure. The London, Edinburgh, and Dublin Philosophical Magazine and
-             Journal of Science, 44(264), 1152-1159.
-
-    .. [#f3] Levine, S., & Neale, G. H. (1975). Theory of the rate of wetting of a
-             porous medium. Journal of the Chemical Society, Faraday Transactions 2:
-             Molecular and Chemical Physics, 71, 12-21.
-    """
-
-    def washburn(t, a, b):
-        return np.sqrt(2 * b / a * t)
-
-    def washburn_rideal(z, a, b):
-        return np.piecewise(
-            z,
-            [z > 0, z == 0],
-            [lambda z: a / 2 / b * z**2 - 1 / a * np.log(a / np.sqrt(b) * z), 0],
-        )
-
-    ret, _ = curve_fit(washburn_rideal, z, t)
-
-    def func(t):
-        t = np.array(t)
-        return root(lambda z: washburn_rideal(z, *ret) - t, washburn(t, *ret)).x
-
-    return func, ret
 
 
 def unidirect_analyzer(k, v):
@@ -216,10 +128,11 @@ def unidirect_analyzer(k, v):
             pip install wettingfront[img]
 
     Unidirectional analyzer detects the horizontal wetting front in the image by
-    pixel intensities and fits the data using :func:`fit_washburn`.
+    pixel intensities and fits the data to model.
 
     The analyzer defines the following fields in configuration entry:
 
+    - **model** (`str`): Wetting front model, implemented by plugins.
     - **path** (`str`): Path to the input video file.
     - **output-vid** (`str`, optional): Path to the output video file.
     - **output-data** (`str`, optional): Path to the output csv file.
@@ -233,13 +146,19 @@ def unidirect_analyzer(k, v):
 
         foo:
             type: Unidirectional
+            model: Washburn
             path: foo.mp4
             output-vid: output/foo.mp4
             output-data: output/foo.csv
     """
     import imageio.v3 as iio
 
+    MODELS = {}
+    for ep in entry_points(group="wettingfront.models"):
+        MODELS[ep.name] = ep
+
     # Prepare output
+    model = MODELS[v["model"]].load()
     path = os.path.expandvars(v["path"])
     out_vid = v.get("output-vid")
     out_data = v.get("output-data")
@@ -280,13 +199,13 @@ def unidirect_analyzer(k, v):
     # write data
     if out_data:
         times = np.arange(len(heights)) / fps
-        func, (k,) = fit_washburn(times, heights)
-        washburn = func(times, k)
+        func, args = model(times, heights)
+        predict = func(times, *args)
 
         with open(out_data, "w", newline="") as f:
             writer = csv.writer(f)
             writer.writerow(["time (s)", "height (pixels)", "fitted height (pixels)"])
-            for t, h, w in zip(times, heights, washburn):
+            for t, h, w in zip(times, heights, predict):
                 writer.writerow([t, h, w])
